+++ mycodo/inputs/examples/minimal_humidity_temperature.py
@@ -33,6 +33,13 @@ INPUT_INFORMATION = {
     'measurements_dict': measurements_dict,
     'measurements_use_same_timestamp': True,
 
+    'url_manufacturer': 'https://www.st.com/en/imaging-and-photonics-solutions/vl53l0x.html',
+    'url_datasheet': 'https://www.st.com/resource/en/datasheet/vl53l0x.pdf',
+    'url_product_purchase': [
+        'https://www.adafruit.com/product/3317',
+        'https://www.pololu.com/product/2490'
+    ],
+
     'message': "Note: This is just an informative note to the user.",
 
     'options_enabled': [